+++ components/mapbox/hooks/layers.js
@@ -1,11 +1,8 @@
 import {useMemo} from 'react'
 
 import {getContourFillLayer, getContourLineLayer} from '../layers/contour'
-import {getVoiesLabelLayer, getToponymesLabelLayer} from '../layers/voies'
-import {getNumerosPointLayer, getNumerosLabelLayer} from '../layers/numeros'
-import {getPositionsPointLayer, getPositionsLabelLayer} from '../layers/positions'
 
-function useLayers(contour, voies, numeros, numero) {
+function useLayers(contour) {
   return useMemo(() => {
     const layers = []
 
@@ -14,23 +11,8 @@ function useLayers(contour, voies, numeros, numero) {
       layers.push(getContourLineLayer())
     }
 
-    if (voies) {
-      layers.push(getVoiesLabelLayer())
-      layers.push(getToponymesLabelLayer())
-    }
-
-    if (numeros) {
-      layers.push(getNumerosPointLayer())
-      layers.push(getNumerosLabelLayer())
-    }
-
-    if (numero) {
-      layers.push(getPositionsPointLayer())
-      layers.push(getPositionsLabelLayer())
-    }
-
     return layers
-  }, [contour, voies, numeros, numero])
+  }, [contour])
 }
 
 export default useLayers
