+++ package.json
@@ -32,7 +32,6 @@
     "randomcolor": "^0.6.2",
     "react": "^17.0.1",
     "react-autocomplete": "^1.7.2",
-    "react-chartjs-2": "^2.11.1",
     "react-dom": "^17.0.1",
     "react-dropzone": "^10.1.10",
     "react-feather": "^2.0.9",
